ImagesTable: Split big files
Split details and instances into separate component files.
