+++ src/Components/ImagesTable/components/ImageDetails/index.tsx
@@ -1,0 +1,6 @@
+export { default as AwsDetails } from './components/AwsDetails';
+export { default as AwsS3Details } from './components/AwsS3Details';
+export { default as AzureDetails } from './components/AzureDetails';
+export { default as GcpDetails } from './components/GcpDetails';
+export { default as LocalDetails } from './components/LocalDetails';
+export { default as OciDetails } from './components/OciDetails';
