+++ src/Components/ImagesTable/components/Instance/index.tsx
@@ -1,0 +1,2 @@
+export { default as AwsS3Instance } from './components/AwsS3Instance';
+export { default as LocalInstance } from './components/LocalInstance';
