+++ src/Components/Launch/GcpLaunchModal.tsx
@@ -29,7 +29,7 @@ import {
   isGcpUploadRequestOptions,
   isGcpUploadStatus,
 } from '../../store/typeGuards';
-import { parseGcpSharedWith } from '../ImagesTable/components/ImageDetails/ImageDetails';
+import { parseGcpSharedWith } from '../ImagesTable/components/ImageDetails/components/GcpDetails';
 
 type LaunchProps = { compose: ComposesResponseItem };
